refactor: split plugins into dedicated folder
