--- packages/vercel/src/universal-middleware.ts
@@ -1,1 +1,0 @@
-export * from "@universal-middleware/vercel";
