+++ packages/vercel/tsup.config.ts
@@ -1,19 +1,18 @@
 import { defineConfig } from "tsup";
 
 const entry = {
-  index: "./src/index.ts",
+  index: "./src/plugins/index.ts",
   utils: "./src/utils/index.ts",
   api: "./src/api.ts",
-  "universal-middleware": "./src/universal-middleware.ts",
 };
 
 export default defineConfig([
   {
     entry,
-    external: ["esbuild", "vike"],
-    format: ["esm", "cjs"],
+    external: ["esbuild"],
+    format: "esm",
     platform: "node",
-    target: "node18",
+    target: "es2022",
     dts: {
       entry,
       compilerOptions: {
